+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Allow choosing the box model in `Node::getPosition()`
 * Allow setting `referrer`, `referrerPolicy` and `transitionType` in `Page::navigate()`
 * Add `Page::authenticate` and `Page::clearAuthentication` methods
+* Add `AbstractBinaryInput::saveToStream` method
 
 
 ## 1.15.1 (UPCOMING)